+++ app/common/__init__.py
@@ -1,5 +1,5 @@
 from common.singleton import Singleton
-from common.util import (
+from common.message_converters import (
     convert_iris_message_to_langchain_message,
     convert_langchain_message_to_iris_message,
 )
